requirements.txt
Flask==1.0.2
bcrypt==3.1.4;
python-dotenv==0.10.0
psycopg2==2.7.6.1
gunicorn==19.9.0